Remove PPOTrainer (#7020)
