+++ docs/source/_toctree.yml
@@ -131,8 +131,6 @@
     title: Online DPO
   - local: orpo_trainer
     title: ORPO
-  - local: ppo_trainer
-    title: PPO
   - local: prm_trainer
     title: PRM
   - local: sdft_trainer
