+++ docs/source/index.md
@@ -27,7 +27,6 @@ Below is the current list of TRL trainers, organized by method type (⚡️ = vL
 - [`RLOOTrainer`](rloo_trainer) ⚡️
 - [`OnlineDPOTrainer`](online_dpo_trainer) 🧪 ⚡️
 - [`NashMDTrainer`](nash_md_trainer) 🧪 ⚡️
-- [`PPOTrainer`](ppo_trainer) 🧪
 - [`XPOTrainer`](xpo_trainer) 🧪 ⚡️
 
 ### Reward modeling
