+++ docs/source/reducing_memory_usage.md
@@ -297,15 +297,6 @@ training_args = OnlineDPOConfig(..., ds3_gather_for_generation=False)
 ```
 
 </hfoption>
-<hfoption id="PPO">
-
-```python
-from trl.experimental.ppo import PPOConfig
-
-training_args = PPOConfig(..., ds3_gather_for_generation=False)
-```
-
-</hfoption>
 <hfoption id="RLOO">
 
 ```python
