--- trl/experimental/ppo/__init__.py
@@ -1,30 +1,0 @@
-# Copyright 2020-2026 The HuggingFace Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from .modeling_value_head import (
-    AutoModelForCausalLMWithValueHead,
-    AutoModelForSeq2SeqLMWithValueHead,
-    PreTrainedModelWrapper,
-)
-from .ppo_config import PPOConfig
-from .ppo_trainer import PPOTrainer
-
-
-__all__ = [
-    "AutoModelForCausalLMWithValueHead",
-    "AutoModelForSeq2SeqLMWithValueHead",
-    "PreTrainedModelWrapper",
-    "PPOConfig",
-    "PPOTrainer",
-]
